Fixed ABS loading instructions.

+++ src/cpu.h
@@ -59,7 +59,7 @@ void reset_cpu(CPU *cpu, MEMORY *memory);
 void initialise_memory(MEMORY *memory);
 int execute_instruction(CPU *cpu, MEMORY *memory);
 BYTE fetch_byte(int *clock, CPU *cpu, MEMORY *memory);
-BYTE read_byte(int *clock, BYTE address, MEMORY *memory);
+BYTE read_byte(int *clock, WORD address, MEMORY *memory);
 void write_byte(int *clock, BYTE address, BYTE value, CPU *cpu, MEMORY *memory);
 WORD fetch_word(int *clock, CPU *cpu, MEMORY *memory);
 void write_word(int *clock, WORD address, WORD value, MEMORY *memory);
